refactor: move webapi execution boundary
