+++ src/image_annotator_lib/webapi/__init__.py
@@ -1,2 +1,1 @@
 """WebAPI execution support package."""
-
